proyectoQr/requirements.txt
Django==4.0.3
Pillow
django-qr-code
numpy
qrcode
cryptography
barcode
pyzbar
opencv-python==4.4.0.46
python-dotenv
barcode==0.13.1